split variants_converter
